requirements.txt
dbt==1.0.0.37.12
dbt-postgres==1.8.0
protobuf<5.0.0,>=4.0.0
pandas==2.0.3
openpyxl==3.1.3
python-dotenv>=0.20.0
psycopg2==2.9.5
requests==2.31.0
tqdm==4.66.4
frictionless==5.17.0